pyproject.toml
[project]
name = "agent-playground"
version = "0.1.0"
description = ""
authors = [
    {name = "Barney Nguyen",email = "huu.thonguyen76@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "gradio (>=5.18.0,<6.0.0)",
    "langgraph (>=0.2.75,<0.3.0)",
    "aisuite[all] (>=0.1.10,<0.2.0)",
    "docstring-parser (>=0.16,<0.17)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pymongo (==4.7.2)",
    "mongoengine (==0.27.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "loguru (==0.6.0)",
    "youtube-transcript-api (>=1.0.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
